pyproject.toml
[project]
name = "bidflow"
version = "0.1.0"
description = "AI-powered RFP Analysis System"
authors = [
    {name = "AI6_6team", email = "team@example.com"},
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "python-multipart>=0.0.6",
    "streamlit>=1.30.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.1",
    "langchain-chroma>=0.1.0",
    "langfuse>=2.0.0",
    "chromadb>=0.4.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.23.0",
    "pdfplumber>=0.10.0",
    "httpx>=0.26.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "olefile>=0.46",
    "pyhwp>=0.1b17",
    "ragas>=0.2.0",
    "datasets>=2.14.0",
    "sentence-transformers>=2.0.0",
    "python-docx>=1.0.0",
    "pyyaml>=6.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
